+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 * `KcFinder` has been completely replaced with `ElFinder` and all its classes
     have been removed. `\MeCms\Command\Install\FixElFinderCommand` and
     `\MeCms\Utility\Checkups\ElFinder` classes have been added;
-* added `\MeCms\Utility\Checkups\Directories` class. `Backups`, `TMP` and
-    `Webroot` checkups classes have been deleded.
+* admin "checkup" function and all its classes have been removed.
 
 ## 2.28 branch
 ### 2.28.1